Moved graphic assets to Resources folder. Changed debug pdb generation to portable format. Moved FiddlerAppContext to separate file. Further .csproj cleanup.
