+++ UoFiddler/Program.cs
@@ -13,55 +13,10 @@ using System;
 using System.IO;
 using System.Windows.Forms;
 using Serilog;
-using UoFiddler.Classes;
-using UoFiddler.Controls.UserControls;
 using UoFiddler.Forms;
 
 namespace UoFiddler
 {
-    internal class FiddlerAppContext : ApplicationContext
-    {
-        private readonly ILogger _logger;
-
-        internal FiddlerAppContext(ILogger logger)
-        {
-            _logger = logger;
-
-            Application.EnableVisualStyles();
-            Application.SetCompatibleTextRenderingDefault(false);
-            Application.ApplicationExit += OnApplicationExit;
-
-            FiddlerOptions.SetLogger(_logger);
-            FiddlerOptions.Startup();
-
-            _logger.Information("Starting loading profile form...");
-            var profile = new LoadProfile { TopMost = true };
-            var profileResult = profile.ShowDialog();
-            if (profileResult == DialogResult.Cancel)
-            {
-                _logger.Information("No profile loaded... exiting.");
-                return;
-            }
-
-            if (FiddlerOptions.UpdateCheckOnStart)
-            {
-                _logger.Information("Update check. Current version is {currentVersion}", FiddlerOptions.AppVersion);
-                UpdateRunner.RunAsync(FiddlerOptions.RepositoryOwner, FiddlerOptions.RepositoryName, FiddlerOptions.AppVersion, false).GetAwaiter().GetResult();
-            }
-
-            _logger.Information("Starting main form...");
-            MainForm = new MainForm();
-            MainForm.Show();
-        }
-
-        private void OnApplicationExit(object sender, EventArgs e)
-        {
-            FiddlerOptions.SaveProfile();
-            Map.SaveMapOverlays();
-            _logger.Information("UOFiddler - Application exit.");
-        }
-    }
-
     internal static class Program
     {
         /// <summary>
